+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fixed an issue where "Share logs" wasn't working.
-- On the Discover tab, center the category buttons.
+- Discover tab: Center the category buttons.
+- Discover tab: Remove placeholder categories and people.
 
 ## [0.1.12] - 2024-05-07Z